Small reorganization to llm unit tests on CPU
