+++ python/llm/test/inference/test_transformers_api.py
@@ -23,7 +23,7 @@ import torch
 import pytest
 
 from bigdl.llm.transformers import AutoModel, AutoModelForCausalLM, AutoModelForSpeechSeq2Seq
-from transformers import AutoTokenizer
+from transformers import AutoTokenizer, LlamaTokenizer
 
 class TestTransformersAPI(unittest.TestCase):
 
@@ -136,6 +136,35 @@ def test_load_low_bit_completion(Model, Tokenizer, model_path, prompt, answer):
             output_str = tokenizer.decode(output[0], skip_special_tokens=True)
 
             assert answer in output_str
+
+prompt = "Once upon a time, there existed a little girl who liked to have adventures. She wanted to go to places and meet new people, and have fun"
+
+@pytest.mark.parametrize("Model, Tokenizer, model_path, prompt", [
+    (AutoModelForCausalLM, LlamaTokenizer, os.environ.get('LLAMA_ORIGIN_PATH'), prompt),
+    (AutoModelForCausalLM, AutoTokenizer, os.environ.get('BLOOM_ORIGIN_PATH'), prompt),
+    (AutoModel, AutoTokenizer, os.environ.get('ORIGINAL_CHATGLM2_6B_PATH'), prompt),
+    (AutoModelForCausalLM, AutoTokenizer, os.environ.get('ORIGINAL_REPLIT_CODE_PATH'), prompt)
+])
+    
+def test_optimize_model(Model, Tokenizer, model_path, prompt):
+    tokenizer = Tokenizer.from_pretrained(model_path, trust_remote_code=True)
+    input_ids = tokenizer.encode(prompt, return_tensors="pt")
+
+    model = Model.from_pretrained(model_path,
+                                load_in_4bit=True,
+                                optimize_model=False,
+                                trust_remote_code=True)
+    logits_base_model = (model(input_ids)).logits
+
+    model = Model.from_pretrained(model_path,
+                                load_in_4bit=True,
+                                optimize_model=True,
+                                trust_remote_code=True)
+    logits_optimized_model = (model(input_ids)).logits
+    diff = abs(logits_base_model - logits_optimized_model).flatten()
+
+    assert any(diff) is False
+
 
 if __name__ == '__main__':
     pytest.main([__file__])
