+++ python/llm/test/inference/test_transformesr_api_434.py
@@ -16,23 +16,18 @@
 
 import os
 import pytest
+import tempfile
 
-from bigdl.llm.transformers import AutoModelForCausalLM, AutoModel
-from transformers import LlamaTokenizer, AutoTokenizer
+from bigdl.llm.transformers import AutoModelForCausalLM
+from transformers import AutoTokenizer
 
 
-llama_model_path = os.environ.get('LLAMA_ORIGIN_PATH')
-bloom_model_path = os.environ.get('BLOOM_ORIGIN_PATH')
-chatglm2_6b_model_path = os.environ.get('ORIGINAL_CHATGLM2_6B_PATH')
-replit_code_model_path = os.environ.get('ORIGINAL_REPLIT_CODE_PATH')
+mistral_model_path = os.environ.get('MISTRAL_ORIGIN_PATH')
 
 prompt = "Once upon a time, there existed a little girl who liked to have adventures. She wanted to go to places and meet new people, and have fun"
 
 @pytest.mark.parametrize("Model, Tokenizer, model_path, prompt", [
-    (AutoModelForCausalLM, LlamaTokenizer, llama_model_path, prompt),
-    (AutoModelForCausalLM, AutoTokenizer, bloom_model_path, prompt),
-    (AutoModel, AutoTokenizer, chatglm2_6b_model_path, prompt),
-    (AutoModelForCausalLM, AutoTokenizer, replit_code_model_path, prompt)
+    (AutoModelForCausalLM, AutoTokenizer, mistral_model_path, prompt)
 ])
     
 def test_optimize_model(Model, Tokenizer, model_path, prompt):
@@ -54,6 +49,31 @@ def test_optimize_model(Model, Tokenizer, model_path, prompt):
 
     assert any(diff) is False
 
+@pytest.mark.parametrize('prompt, answer', [
+    ('What is the capital of France?\n\n', 'Paris')
+    ])
+@pytest.mark.parametrize('Model, Tokenizer, model_path',[
+    (AutoModelForCausalLM, AutoTokenizer, mistral_model_path),
+    ])
+def test_load_low_bit_completion(Model, Tokenizer, model_path, prompt, answer):
+    tokenizer = Tokenizer.from_pretrained(model_path, trust_remote_code=True)
+    model = Model.from_pretrained(model_path,
+                                  load_in_4bit=True,
+                                  optimize_model=True,
+                                  trust_remote_code=True)
+    
+    with tempfile.TemporaryDirectory() as tempdir:
+        model.save_low_bit(tempdir)
+        loaded_model = Model.load_low_bit(tempdir,
+                                          optimize_model=True,
+                                          trust_remote_code=True)
+
+        with torch.inference_mode():
+            input_ids = tokenizer.encode(prompt, return_tensors="pt")
+            output = loaded_model.generate(input_ids, max_new_tokens=32)
+            output_str = tokenizer.decode(output[0], skip_special_tokens=True)
+
+            assert answer in output_str
 
 if __name__ == '__main__':
     pytest.main([__file__])
