+++ python/llm/test/run-llm-inference-tests.sh
@@ -9,18 +9,17 @@ set -e
 echo "# Start testing inference"
 start=$(date "+%s")
 
-python -m pytest -s ${LLM_INFERENCE_TEST_DIR} -k "not test_transformers" -v \
-      --ignore=${LLM_INFERENCE_TEST_DIR}/test_optimize_mistral.py
+python -m pytest -s ${LLM_INFERENCE_TEST_DIR}/test_call_models.py -v
 
 if [ -z "$THREAD_NUM" ]; then
   THREAD_NUM=2
 fi
 export OMP_NUM_THREADS=$THREAD_NUM
-python -m pytest -s ${LLM_INFERENCE_TEST_DIR} -k test_transformers -v \
-      --ignore=${LLM_INFERENCE_TEST_DIR}/test_optimize_mistral.py
+python -m pytest -s ${LLM_INFERENCE_TEST_DIR}/test_transformers_api.py -v
+python -m pytest -s ${LLM_INFERENCE_TEST_DIR}/test_optimize_model_api.py -v
 
 python -m pip install transformers==4.34.0
-python -m pytest -s ${LLM_INFERENCE_TEST_DIR}/test_optimize_mistral.py -v
+python -m pytest -s ${LLM_INFERENCE_TEST_DIR}/test_transformesr_api_434.py -v
 python -m pip install transformers==4.31.0
 
 now=$(date "+%s")
